refactor comboboxfieldstate into codedvaluefieldstate
